# AlamiaAccounts-Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AlamiaAccounts API",
        "description": "API testing",
        "version": "1.0.0"
    },
    "paths": {
...
        "/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get all companies",
                "operationId": "3da78f5f344ed7d6004d63537a662220",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "COMP1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Company One"
                                                    },
                                                    "industry": {
                                                        "type": "string",
                                                        "example": "General"
                                                    },
                                                    "currency": {
                                                        "type": "string",
                                                        "example": "USD"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "operationId": "da9e9a653ddfa3aa276a6dd441ce5baa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "industry",
                                    "currency"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Company Name"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "NEWCO"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Technology"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "NEWCO"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "New Company Name"
                                                },
                                                "industry": {
                                                    "type": "string",
                                                    "example": "Technology"
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "USD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Failed to create company"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/companies/{code}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get company details",
                "operationId": "93c0816fc7a0ec66b8d59adb1ac892f2",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "COMP1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "COMP1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Company One"
                                                },
                                                "industry": {
                                                    "type": "string",
                                                    "example": "General"
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "USD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update an existing company",
                "operationId": "7dc6dd45af248940abc7b2e0b6468527",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "COMP1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "industry"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Company Name"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Finance"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "COMP1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Updated Company Name"
                                                },
                                                "industry": {
                                                    "type": "string",
                                                    "example": "Finance"
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "USD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "operationId": "ec01f741366ad4f981597dfbf01e5040",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "COMP1"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Company deleted successfully"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/companies/{code}/switch": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Switch current company context",
                "operationId": "5fab0508d0e7caf171561c8197ee81f0",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "COMP1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company context switched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Company switched successfully"
                                        },
                                        "current_company": {
                                            "type": "string",
                                            "example": "COMP1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}